+++ R/wflow_paths.R
@@ -27,8 +27,15 @@ wflow_paths <- function(error_git = FALSE, project = ".") {
                 silent = TRUE)
   if (class(o$root) == "try-error")
     stop(wrap(
-      "Unable to find RStudio Rproj file at the root of the workflowr project.
-      Did you delete it?"),
+      "Unable to detect a workflowr project. This could be due to one of the
+      following reasons:
+
+      1) The function was not executed inside a workflowr project. Run
+      `getwd()` to determine the current working directory. Is the working
+      directory a workflowr project or one of its subdirectories?
+
+      2) The RStudio .Rproj file was deleted. workflowr requires an RStudio
+      .Rproj file to be located at the root of the project. Was it deleted?"),
       call. = FALSE)
 
   # Analysis directory with _site.yml