+++ NEWS.md
@@ -1,5 +1,7 @@
 # workflowr 0.7.0.9000
 
+* Minor changes to improve error messages and documentation.
+
 * Added tag to footer.html to adjust MathJax settings so that all math
 formulae are shown using TeX fonts only. This will make the
 presentation more consistent at the cost of the webpage sometimes